+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [#1061](https://github.com/crypto-org-chain/chain-main/pull/1061) Integrate sdk 0.50.
 - [#1068](https://github.com/crypto-org-chain/chain-main/pull/1068) Upgrade ibc-go to `v8.3.2` and remove icaauth module.
 - [#1084](https://github.com/crypto-org-chain/chain-main/pull/1084) Add MsgModuleQuerySafe in allowed messages for ica host param.
+- [#1085](https://github.com/crypto-org-chain/chain-main/pull/1085) Upgrade ibc-go to `v8.5.1`.
 
 *Dec 6, 2023*